nyayai-backend/requirements.txt
fastapi
uvicorn
pydantic
llama-cpp-python
onnxruntime
pytesseract
pdfplumber
Pillow
sentence-transformers
langchain
chromadb
sqlite3-binary
indic-transliteration
transformers
torch
